refactor: Update measurement-guide directory
