+++ labware-library/src/components/measurement-guide/images/depth/depth-reservoir-and-tubes-flat@3x.png
