+++ labware-library/src/components/measurement-guide/images/depth/depth-reservoir-and-tubes-round@3x.png
