+++ labware-library/src/components/measurement-guide/images/dimensions/footprint@3x.png
