+++ labware-library/src/components/measurement-guide/images/offset/offset-reservoir@3x.png
