+++ labware-library/src/components/measurement-guide/images/offset/offset-well-circular@3x.png
