+++ labware-library/src/components/measurement-guide/images/offset/offset-well-rectangular@3x.png
