+++ labware-library/src/components/measurement-guide/images/shape/shape-circular@3x.png
