+++ labware-library/src/components/measurement-guide/images/shape/shape-rectangular@3x.png
