+++ labware-library/src/components/measurement-guide/images/spacing/spacing-reservoir@3x.png
