+++ labware-library/src/components/measurement-guide/images/spacing/spacing-well-circular@3x.png
