+++ labware-library/src/components/measurement-guide/images/spacing/spacing-well-rectangular@3x.png
